Refactored some tests
